# Documents/LeFax Talents Dev Projects /backendsempos-master/api-doc/apidoc_v1.yaml
openapi: 3.0.2
info:
  title: Onebaz API
  version: '1.0'
  contact:
    name: Onebaz
    email: sempos@gmx.com
    url: www.onebaz.com
  description: API for the Onebaz back-end system
servers:
  - url: 'https://api.onebaz.com/v1/base'
  - url: 'https://api.onebaz.com/v1/{tenant}'
    variables:
      tenant:
        description: Tenant name
        default: lefax
        enum:
          - lefax
          - yellowshop
          - salaudbar
          - idealdallas
  - url: 'http://localhost:3000/v1/base'
  - url: 'http://localhost:3000/v1/{tenant}'
    variables:
      tenant:
        description: Tenant name
        default: velocity
        enum:
          - velocity
          - yellowshop
          - salaudbar
          - idealdallas
paths:
...
  /transfers:
    get:
      summary: Get transfers
      tags:
        - Transfers
      responses:
        '200':
          $ref: '#/components/responses/GetTransfers'
      operationId: get-transfers
      security:
        - Tenant User Api Key: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                transferred_by:
                  type: array
                  items:
                    type: string
                transfer_to:
                  type: array
                  items:
                    type: string
                status:
                  type: array
                  items:
                    type: string
                start_date:
                  type: string
                end_date:
                  type: string
    post:
      summary: Add a new transfer
      operationId: post-transfers
      responses:
        '201':
          $ref: '#/components/responses/GetTransfer'
      security:
        - Tenant User Api Key: []
      requestBody:
        $ref: '#/components/requestBodies/AddTransfer'
      tags:
        - Transfers
  '/transfers/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Get a transfer by id
      tags:
        - Transfers
      responses:
        '200':
          $ref: '#/components/responses/GetTransfer'
      operationId: get-transfers-id
      security:
        - Tenant User Api Key: []
    patch:
      summary: Update a transfer
      operationId: patch-transfers-id
      responses:
        '200':
          $ref: '#/components/responses/GetTransfer'
      security:
        - Tenant User Api Key: []
      requestBody:
        $ref: '#/components/requestBodies/UpdateTransfer'
      tags:
        - Transfers
  /transfers/verify:
    get:
      summary: Check the availability of a transfer number
      tags:
        - Transfers
      operationId: get-transfers-verify
      security:
        - Tenant User Api Key: []
      parameters:
        - schema:
            type: string
          in: query
          name: transfer_number
          required: true
      responses:
        '200':
          $ref: '#/components/responses/ResponseMessage'
  /transaction_categories:
    get:
      summary: Get treasury transaction categories
      tags:
        - Treasury Transaction Category
      responses:
        '200':
          $ref: '#/components/responses/GetTreasuryTransactionCategories'
      operationId: get-transaction_categories
      security:
        - Tenant User Api Key: []
      parameters:
        - schema:
            type: string
          in: query
          name: type
    post:
      summary: Add a treasury transaction category
      operationId: post-transaction_categories
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetTreasuryTransactionCategory'
                  url:
                    type: string
                    format: uri
      security:
        - Tenant User Api Key: []
      requestBody:
        $ref: '#/components/requestBodies/AddTreasuryTransactionCategory'
      tags:
        - Treasury Transaction Category
  '/transaction_categories/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Update a treasury transaction category
      tags:
        - Treasury Transaction Category
      responses:
        '200':
          $ref: '#/components/responses/GetTreasuryTransactionCategory'
      operationId: get-transaction_categories-id
      security:
        - Tenant User Api Key: []
    patch:
      summary: Update a treasury transaction category
      operationId: patch-transaction_categories-id
      responses:
        '200':
          $ref: '#/components/responses/GetTreasuryTransactionCategories'
      security:
        - Tenant User Api Key: []
      requestBody:
        $ref: '#/components/requestBodies/UpdateTreasuryTransactionsCategory'
      tags:
        - Treasury Transaction Category
    delete:
      summary: Delete a treasury transaction category
      operationId: delete-transaction_categories-id
      responses:
        '200':
          $ref: '#/components/responses/ResponseMessage'
      security:
        - Tenant User Api Key: []
      tags:
        - Treasury Transaction Category
  /transaction_categories/verify:
    get:
      summary: Check the availability of a category name
      tags:
        - Treasury Transaction Category
      responses:
        '200':
          $ref: '#/components/responses/ResponseMessage'
      operationId: get-transaction_categories-verify
      security:
        - Tenant User Api Key: []
      parameters:
        - schema:
            type: string
          in: query
          name: name
          required: true
  /transactions:
    get:
      summary: Get transactions
      tags:
        - Cash Flow Transactions
      responses:
        '200':
          $ref: '#/components/responses/GetCashFlowTransactions'
      operationId: get-transactions
      security:
        - Tenant User Api Key: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                types:
                  type: array
                  uniqueItems: true
                  minItems: 0
                  maxItems: 2
                  items:
                    type: string
                    enum:
                      - collection
                      - disbursement
                min_amount:
                  type: number
                max_amount:
                  type: integer
                min_tax:
                  type: integer
                max_tax:
                  type: number
                start_billing_date:
                  type: string
                  format: date
                end_billing_date:
                  type: string
                  format: date
                start_payment_date:
                  type: string
                  format: date
                end_payment_date:
                  type: string
                  format: date
                status:
                  type: array
                  uniqueItems: true
                  minItems: 0
                  maxItems: 2
                  items:
                    type: string
                    enum:
                      - underway
                      - completed
                categories:
                  type: array
                  uniqueItems: true
                  items:
                    type: string
    post:
      summary: Add a new transaction
      operationId: post-transactions
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetCashFlowTransaction'
                  url:
                    type: string
                    format: uri
      security:
        - Tenant User Api Key: []
      requestBody:
        $ref: '#/components/requestBodies/AddCashFlowTransaction'
      tags:
        - Cash Flow Transactions
  '/transactions/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Get a transaction by id
      tags:
        - Cash Flow Transactions
      responses:
        '200':
          $ref: '#/components/responses/GetCashFlowTransaction'
      operationId: get-transactions-id
      security:
        - Tenant User Api Key: []
    patch:
      summary: Update a transaction
      operationId: patch-transactions-id
      responses:
        '200':
          $ref: '#/components/responses/GetCashFlowTransaction'
      security:
        - Tenant User Api Key: []
      requestBody:
        $ref: '#/components/requestBodies/UpdateCashFlowTransaction'
      tags:
        - Cash Flow Transactions
    delete:
      summary: Delete a transaction
      operationId: delete-transactions-id
      responses:
        '200':
          $ref: '#/components/responses/ResponseMessage'
      security:
        - Tenant User Api Key: []
      tags:
        - Cash Flow Transactions
  /transactions/verify:
    get:
      summary: Check the availability of a transaction name
      tags:
        - Cash Flow Transactions
      responses:
        '200':
          $ref: '#/components/responses/ResponseMessage'
      operationId: get-transactions-verify
      security:
        - Tenant User Api Key: []
      parameters:
        - $ref: '#/components/parameters/name'
...
components:
  schemas:
...
    AddTransfer:
      title: AddTransfer
      type: object
      properties:
        details:
          type: object
          properties:
            transfer_number:
              type: string
            issue_date:
              type: string
              format: date-time
            expected_transfer_date:
              type: string
              format: date-time
            branch_source:
              type: string
            branch_target:
              type: string
            note:
              type: string
          required:
            - expected_transfer_date
            - branch_source
        products:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              quantities:
                type: object
                properties:
                  requested:
                    type: number
                required:
                  - requested
            required:
              - id
    GetTransfer:
      title: GetTransfer
      type: object
      properties:
        tenant:
          type: string
        employee:
          type: object
          properties:
            id:
              type: string
            firstname:
              type: string
            lastname:
              type: string
        details:
          type: object
          properties:
            transfer_number:
              type: string
            issue_date:
              type: string
              format: date-time
            expected_transfer_date:
              type: string
              format: date-time
            branch_source:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
            branch_target:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
            status:
              type: string
              default: waiting_for_approval
              enum:
                - waiting_for_approval
                - waiting_transfer
                - completed
            note:
              type: string
        products:
          type: array
          items:
            type: object
            properties:
              product:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
              status:
                type: string
                default: waiting
                enum:
                  - waiting
                  - received
                  - not_received
                  - distinct
                  - damaged
              quantities:
                type: object
                properties:
                  requested:
                    type: number
                  received:
                    type: number
                  not_received:
                    type: number
                  damaged:
                    type: number
    UpdateTransfer:
      title: UpdateTransfer
      type: object
      properties:
        details:
          type: object
          properties:
            transfer_number:
              type: string
            issue_date:
              type: string
              format: date-time
            expected_transfer_date:
              type: string
              format: date-time
            branch_source:
              type: string
            branch_target:
              type: string
            status:
              type: string
              default: waiting_for_approval
              enum:
                - waiting_for_approval
                - waiting_transfer
                - completed
            note:
              type: string
          required:
            - branch_source
            - branch_target
            - status
        products:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              status:
                type: string
                default: waiting
                enum:
                  - waiting
                  - received
                  - not_received
                  - distinct
                  - damaged
              quantities:
                type: object
                properties:
                  requested:
                    type: number
                  received:
                    type: number
                  not_received:
                    type: number
                  damaged:
                    type: number
    AddTreasuryTransactionCategory:
      title: AddTreasuryTransactionCategory
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - collection
            - disbursement
      required:
        - name
        - type
    GetTreasuryTransactionCategory:
      title: GetTreasuryTransactionCategory
      type: object
      properties:
        id:
          type: string
        tenant:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - collection
            - disbursement
    UpdateTreasuryTransactionCategory:
      title: UpdateTreasuryTransactionCategory
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - collection
            - disbursement
      required:
        - name
        - type
    GetCashFlowTransaction:
      title: GetCashFlowTransaction
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - collection
            - disbursment
        amount:
          type: number
        tax:
          type: number
        billing_date:
          type: string
          format: date-time
        payment_date:
          type: string
          format: date
        status:
          type: string
          enum:
            - underway
            - completed
        overdue:
          type: boolean
        category:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
        note:
          type: string
    AddCashFlowTransaction:
      title: AddCashFlowTransaction
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - collection
            - disbursment
        amount:
          type: number
        tax:
          type: number
        billing_date:
          type: string
          format: date-time
        payment_date:
          type: string
          format: date
        status:
          type: string
          enum:
            - underway
            - completed
        category:
          type: string
        note:
          type: string
      required:
        - name
        - type
        - amount
        - category
    UpdateCashFlowTransaction:
      title: UpdateCashFlowTransaction
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - collection
            - disbursment
        amount:
          type: number
        tax:
          type: number
        billing_date:
          type: string
          format: date-time
        payment_date:
          type: string
          format: date
        status:
          type: string
          enum:
            - underway
            - completed
        category:
          type: string
        note:
          type: string
  securitySchemes:
    Admin Api Key:
      name: admin_api_key
      type: apiKey
      in: header
    Tenant User Api Key:
      name: tenant_user_api_key
      type: apiKey
      in: header
  requestBodies:
...
    AddTransfer:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AddTransfer'
    UpdateTransfer:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateTransfer'
    AddTreasuryTransactionCategory:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AddTreasuryTransactionCategory'
    UpdateTreasuryTransactionsCategory:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateTreasuryTransactionCategory'
    AddCashFlowTransaction:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AddCashFlowTransaction'
        application/xml:
          schema:
            type: object
            properties: {}
        multipart/form-data:
          schema:
            type: object
            properties: {}
    UpdateCashFlowTransaction:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateCashFlowTransaction'
  parameters:
...
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: string
    name:
      name: name
      in: query
      required: true
      schema:
        type: string
  responses:
...
    ResponseMessage:
      description: Example response
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
...
    GetTransfer:
      description: Example response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetTransfer'
    GetTransfers:
      description: Example response
      content:
        application/json:
          schema:
            type: object
            properties:
              total:
                type: integer
              count:
                type: integer
              currentPage:
                type: integer
              pages:
                type: integer
              prev:
                type: string
                format: uri
                nullable: true
              next:
                type: string
                format: uri
                nullable: true
              data:
                type: array
                items:
                  $ref: '#/components/schemas/GetTransfer'
    GetTreasuryTransactionCategory:
      description: Example response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetTreasuryTransactionCategory'
    GetTreasuryTransactionCategories:
      description: Example response
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/GetTreasuryTransactionCategory'
    GetCashFlowTransaction:
      description: Example response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GetCashFlowTransaction'
    GetCashFlowTransactions:
      description: Example response
      content:
        application/json:
          schema:
            type: object
            properties:
              total:
                type: integer
              count:
                type: integer
              currentPage:
                type: integer
              pages:
                type: integer
              prev:
                type: string
                format: uri
                nullable: true
              next:
                type: string
                format: uri
                nullable: true
              data:
                type: array
                items:
                  $ref: '#/components/schemas/GetCashFlowTransaction'
...